新しいコンポーネントHistoryTab、MemberTab、SearchSection、TabManagerを追加し、UserSelectStepコンポーネントをリファクタリングして、タブ管理と検索機能を統合。これにより、コードの可読性と再利用性が向上。
